+++ src/api/alarm/types.ts
@@ -1,0 +1,7 @@
+export interface AlarmType {
+  content: string;
+  createdAt: string;
+  isRead: 'N' | 'Y';
+  notificationId: number;
+  postId: number;
+}
