+++ src/api/application/types.ts
@@ -1,0 +1,7 @@
+export interface IApplicantDetail {
+  nickname: string;
+  profileImg: string;
+  email: string;
+  message: string;
+  applicationDate: string;
+}
