+++ src/api/reply/types.ts
@@ -1,0 +1,4 @@
+export interface IRequestReply {
+  commentId: number;
+  content: string;
+}
