+++ src/api/user/types.ts
@@ -1,0 +1,10 @@
+export interface IUserType {
+  email: string;
+  intro: string;
+  nickname: string;
+  position: string;
+  profileImageUrl: string;
+  roleType: string;
+  mentoringNum: number;
+  studyNum: number;
+}
