+++ src/components/domains/createStudy/Buttons/index.tsx
@@ -1,11 +1,10 @@
 import { AxiosResponse } from 'axios';
 import Swal from 'sweetalert2';
-
-import { PostRequestDto } from 'types/api.types';
 import { useCustomNavigate } from '@hooks/useCustomNavigate';
 import { useAddPost, useUpdatePost } from '@hooks/query/usePostQuery';
 import { validateFormData } from '@utils/validateData';
 import { PostAPI } from '@api/post';
+import { PostRequestDto } from '@api/post/types';
 
 interface SubmitButtonProps {
   className: string;
