+++ src/components/domains/detail/CommentCountView/CommentCountView.tsx
@@ -1,10 +1,9 @@
 import { Icon } from '@components/atoms';
-import { IResponsePostDetail } from 'types/api.types';
-
 import HeartIcon from '@assets/svg/common/heart.svg';
 import EyeIcon from '@assets/svg/common/eye.svg';
 import MessageIcon from '@assets/svg/common/message-square.svg';
 import classes from './CommentCountView.module.scss';
+import { IResponsePostDetail } from '@api/post/types';
 
 export const CommentCountView = (
   props: Pick<IResponsePostDetail, 'likeNum' | 'commentNum' | 'view'>,
