+++ src/components/domains/myPage/common/CardsSection/MemberCardsSection/index.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import Pagination from '../../Pagination';
 import classes from './index.module.scss';
 import profileCircle from '../../../../../../assets/svg/profile-circle.svg';
-import { IMemberType } from 'types/member.types';
+import { IMemberType } from '@api/team/types';
 
 interface IMemberCardsSectionProps {
   members?: IMemberType[];
