+++ src/components/domains/myPage/common/CardsSection/MentoringPostCardsSection/index.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import Pagination from '../../Pagination';
 import classes from './index.module.scss';
 import profileCircle from '../../../../../../assets/svg/profile-circle.svg';
-import { IMentoringPostsType } from 'types/posts.types';
+import { IMentoringPostsType } from '@api/post/types';
 
 interface IMentoringPostCardsSectionProps {
   mentoringPosts?: IMentoringPostsType;
