+++ src/store/search.tsx
@@ -1,5 +1,5 @@
+import { ISearchType } from '@api/post/types';
 import { atom } from 'recoil';
-import { ISearchType } from 'types/api.types';
 
 export const searchState = atom<ISearchType>({
   key: 'searchState',
